chatbot_project/requirements.txt
pandas
dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
openai 